Remove unnecessary changes and add initializations.

+++ src/creature_instances.h
@@ -102,7 +102,7 @@ struct InstanceInfo {
     long reset_time;
     long fp_reset_time;
     unsigned char graphics_idx;
-    unsigned long instance_property_flags;
+    short instance_property_flags;
     short force_visibility;
     unsigned char primary_target;
     unsigned char func_idx;
